add delete method

+++ src/main/java/pl/roslon/WindsurfingWindFinder/service/WindService.java
@@ -1,6 +1,7 @@
 package pl.roslon.WindsurfingWindFinder.service;
 
 import jakarta.transaction.Transactional;
+import jakarta.validation.constraints.NotBlank;
 import org.springframework.stereotype.Service;
 import pl.roslon.WindsurfingWindFinder.model.PointDto;
 import pl.roslon.WindsurfingWindFinder.repository.PointRepository;
@@ -17,17 +18,19 @@ public class WindService {
         this.pointRepository = pointRepository;
     }
 
-
-    public void addPointToDb(PointDto pointDto){
+    public void addPointToDb(PointDto pointDto) {
         pointRepository.save(pointDto);
     }
 
-    public List<PointDto> printSortedPointList(){
-       return pointRepository.findAllByOrderByAvgWindSpeedDesc();
+    public List<PointDto> printSortedPointList() {
+        return pointRepository.findAllByOrderByAvgWindSpeedDesc();
     }
 
     @Transactional
-    public void deletePoint(String city) {
+    public void deletePoint(@NotBlank String city) throws IllegalArgumentException {
+        if (!pointRepository.existsByCity(city)) {
+            throw new IllegalArgumentException("Point with city " + city + " does not exist.");
+        }
         pointRepository.deleteByCity(city);
     }
